fix: use middleware from common middleware package
